rpi/fastapi-sample-app/requirements.txt
FastAPI
uvicorn
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi 
opentelemetry-exporter-otlp